Добавил асинхронную обработку водяного знака через BackgroundTasks

+++ app/services/client.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 import uuid
-from fastapi import UploadFile
+from fastapi import UploadFile, BackgroundTasks
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
@@ -10,19 +10,26 @@ from schemas import client
 from PIL import Image
 
 
+def add_watermark(image_path: str, watermark_path: str):
+    original_image = Image.open(image_path)
+    watermark = Image.open(watermark_path)
+    if original_image.mode == 'RGBA':
+        original_image = original_image.convert('RGB')
+    original_image.paste(watermark, (0, 0), watermark)
+    original_image.save(image_path, format='JPEG')
+
+
 def create_client(
     data: client.Client,
     db: Session,
-    profile_pic: UploadFile
+    profile_pic: UploadFile,
+    background_tasks: BackgroundTasks,
 ):
     profile_pic.filename = f"{uuid.uuid4().hex}_{profile_pic.filename.lower()}"
-    path = os.path.join('static', profile_pic.filename)
-    with open(path, 'wb+') as buffer:
+    path = os.path.join("static", profile_pic.filename)
+    with open(path, "wb+") as buffer:
         shutil.copyfileobj(profile_pic.file, buffer)
-    original_image = Image.open(path)
-    watermark = Image.open('static/watermark.png').convert("RGBA")
-    original_image.paste(watermark, (0, 0), watermark)
-    original_image.save(path)
+    background_tasks.add_task(add_watermark, path, "static/watermark.png")
     client = Client(
         name=data.name,
         last_name=data.last_name,
@@ -30,14 +37,14 @@ def create_client(
         mail=data.mail,
         profile_pic=path,
         latitude=data.latitude,
-        longitude=data.longitude
+        longitude=data.longitude,
     )
     try:
         db.add(client)
         db.commit()
         db.refresh(client)
     except Exception as e:
-        print('Ошибка добавления клиента:', e)
+        print("Ошибка добавления клиента:", e)
         raise e
     return client
 
@@ -46,7 +53,13 @@ def get_client(id: int, db):
     return db.query(Client).filter(Client.id == id).first()
 
 
-def update(profile_pic: UploadFile, id: int, data: client.Client, db: Session):
+def update(
+    profile_pic: UploadFile,
+    id: int,
+    data: client.Client,
+    db: Session,
+    background_tasks: BackgroundTasks,
+):
     client = db.query(Client).filter(Client.id == id).first()
     if not client:
         return {"error": "Клиент не найден."}
@@ -57,24 +70,31 @@ def update(profile_pic: UploadFile, id: int, data: client.Client, db: Session):
     client.latitude = data.latitude
     client.longitude = data.longitude
     if profile_pic:
-        os.remove(client.profile_pic)
+        if os.path.exists(client.profile_pic):
+            os.remove(client.profile_pic)
         profile_pic.filename = (
-                            f"{uuid.uuid4().hex}{profile_pic.filename.lower()}"
-                        )
-        path = os.path.join('static', profile_pic.filename)
-        with open(path, 'wb+') as buffer:
-            shutil.copyfileobj(profile_pic.file, buffer)
-        original_image = Image.open(path)
-        watermark = Image.open('static/watermark.png').convert("RGBA")
-        original_image.paste(watermark, (0, 0), watermark)
-        original_image.save(path)
-        client.profile_pic = path
+            f"{uuid.uuid4().hex}"
+            f"{profile_pic.filename.lower()}"
+        )
+        path = os.path.join("static", profile_pic.filename)
+        try:
+            with open(path, "wb+") as buffer:
+                shutil.copyfileobj(profile_pic.file, buffer)
+            background_tasks.add_task(
+                add_watermark,
+                path,
+                "static/watermark.png"
+            )
+            client.profile_pic = path
+        except Exception as e:
+            print("Ошибка при сохранении профиля:", e)
+            return {"error": "Не удалось сохранить изображение профиля."}
     try:
         db.commit()
         db.refresh(client)
     except IntegrityError as e:
         db.rollback()
-        print('Ошибка обновления клиента:', e)
+        print("Ошибка обновления клиента:", e)
         return {"error": "Произошла ошибка при обновлении клиента."}
     return client
